Added qnx_net and qnx_savegame to build. Added linking to library libscreen. No longer excluding RenderMatrix from compiling. Made sure doomclassic's Precompiled doesn't redefine ID_INLINE Changed USE_OSBUILTIN_ZLIB to USE_OS_ZLIB, which is more accurate Changed USE_OSBUILTIN_LIBJPEG to USE_OS_LIBJPEG, which is more accurate Commented out ID_INLINE_EXTERN and ID_FORCE_INLINE_EXTERN and replaced with copies of ID_INLINE and ID_FORCE_INLINE as qnu_inline is an ignored attribute Commented out missing fp_* functions from qnx_cpu Fixed use of winUserState_t instead of qnxUserState_t Check application state when determining if system UI is showing (Temporary for now) Renamed win* variables in signin to qnx Made sure correct local user is used with idSaveLoadParams::SetDefaults Made sure correct system-specific includes are used in sys_session_local.h

+++ doomclassic/doom/i_sound_qnx.cpp
@@ -864,7 +864,7 @@ unsigned int WINAPI I_LoadSong( void* songname ) {
 	int length = 0;
 	Mus2Midi( musFile, midiConversionBuffer, &length );
 
-	doomMusic = Timidity_LoadSongMem( midiConversionBuffer, length );
+	doomMusic = Timidity_LoadSongMem( midiConversionBuffer, ( size_t )length );
 
 	if ( doomMusic ) {
 		musicBuffer = (byte *)malloc( MIDI_CHANNELS * MIDI_FORMAT_BYTES * doomMusic->samples );
